requirements.txt
asgiref==3.2.5
Django==3.0.4
django-admin==2.0.1
django-excel-base==1.0.3
django-excel-response2==3.0.1
django-six==1.0.4
pytz==2019.3
screen==1.0.1
sqlparse==0.3.1
xlwt==1.3.0
Pillow==7.0.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.0